Adding new lending pair data type
