+++ src/data/constants/Addresses.ts
@@ -7,4 +7,5 @@ export const BLEND_FACTORY_CREATION_BLOCK = 14128969;
 export const WETH_9_MAINNET_ADDRESS =
   '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2';
 
-export const ALOE_II_FACTORY_ADDRESS_GOERLI = '0xC49D44f7e3Eb918A9E038EEbB007Ae0292BE5576';
+export const ALOE_II_FACTORY_ADDRESS_GOERLI = '0xc49d44f7e3eb918a9e038eebb007ae0292be5576';
+export const ALOE_II_KITTY_LENS_ADDRESS = '0x723bfe564661536fdffa3e9e060135928d3bf18f';
